# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Baraka API",
        "description": "API backend pour la plateforme Baraka de microcrédit",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/baraka-backend/public",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/auth/otp/request": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Demander un OTP par SMS",
                "operationId": "6e74ccbaecf96c72cdb4b860b7118062",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone"
                                ],
                                "properties": {
                                    "telephone": {
                                        "type": "string",
                                        "example": "+22370000000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP envoyé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/auth/otp/verify": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Vérifier l'OTP et obtenir un token JWT",
                "operationId": "1b1c3480978f35a7bb0eee7fe8aa3bb8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone",
                                    "code"
                                ],
                                "properties": {
                                    "telephone": {
                                        "type": "string",
                                        "example": "+22370000000"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token JWT retourné",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                                },
                                                "user": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Code invalide ou expiré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Code invalide ou expiré."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez le token JWT obtenu via POST /api/auth/otp/verify"
            }
        }
    },
...
}